Experimental support for setting the ellipsoids
There is a global ellipsoid exposed through the `get_ellipsoid`
function. All functions should use that to get the ellipsoid.
Set another ellipsoid using `set_ellipsoid` which also works as a
context manager. Can take an ellipsoid name or an `Ellipsoid` instance.
So the user can set their own ellipsoid and make it permanent for the
whole script or only for a given context.