Global gravity sample dataset (#12)
Added a global gravity sample dataset as a netCDF file downloaded from ICGEM Computation
Service. The data contains the EIGEN 6C4 release gravity_earth field which is defined as
the magnitude of the gradient of the potential (including the centrifugal potential) at
a given point. It was calculated on the Earth's surface and on a regular global grid of
1.0° spacing.