Refactor load_icgem_gdf and simplify coordinates (#16)
The latitude and longitude were being read as 2D arrays and referenced
to "northing" and "easting" helper coordinates. Since this grid is not
projected, there is no need for that.

Rename coordinates to "latitude" and "longitude" instead of "lon" "lat".
Refactor the function to be a bit smaller when creating the array.
Mainly, prepare all arguments before instantiating the Dataset instead
of assigning the values afterward.