Use original ETOPO1 heights and cast ints to float
ETOPO1 is referenced to sea level. I had used a geoid model to convert
it to geometric heights but this is not a great idea. Revert back to the
original. The data were stored as ints to save space but this can cause
problems with divisions and other operations. Cast them back to float
after loading with our functions.

Fixes #29
Fixes #30