Simplify Airy code and make compatible with xarray (#36)
I simplified the implementation of the Airy isostatic Moho to only handle the
traditional cases: crust, mantle, and water density. Don't ignore the water by default.
This can be achieved by masking the topography if desired. Also include the reference
Moho depth in the arguments to output Moho depth instead of root thickness. This is the
most common use case and the roots can still be calculated using 0 reference depth. The
implementation now handles `xarray.DataArray` input and fills out metadata.

Extended the documentation and gallery example text and added a figure to explain the
geometry of the problem.