Function for planar Bouguer correction (#38)
Adds a `bouguer_correction` function that calculate the planar Bouguer correction for
land and oceans. The topography provided can be a numpy array or `xarray.DataArray`.
Includes an example application to the global gravity sample data. In this case, need to
convert the geoid-referenced ETOPO1 topography to ellipsoid heights.

Fixes #8