Move tesseroid checks outside jitted functions (#87)
Refactor the functions that checks if tesseroids are well defined and no computation
points are inside tesseroids. Instead of being Numba jitted, they now make use of Numpy
array properties. The sanity checks are done on the public Python function that performs
the forward modelling instead of calling them on the Numba jitted functions so we can
control the error messages better and disable checks if more speed is desired. Make error
messages more explicit by showing which tesseroids or computation points are not valid.
Improve docstrings by specifying the shape of coordinates and tesseroids arrays. Extend
test functions for computation points inside tesseroids in case the longitudinal
coordinates of points are shifted by 360 degrees.

Fixes #84