Add capability to compute sparse Jacobian matrix
If distance_threshold is not None, the Jacobian matrix will be computed
as a sparse matrix. Only the elements that involve observation and
source points that are closer than the distance_threshold will be taken
as non zero. The indices of the sets of observation points that are
close enough to each point source are computed through the
query_ball_tree method of cKDTree.