Add gravity station data for South Africa (#99)
Add compressed version of the ASCII file that contains the observed
gravity on each station within the boundaries of the Republic of South
Africa and a function to load it as a pandas.DataFrame. Also add a test
function that performs sanity checks on the loaded dataset and an
gallery example for the new feature.