Modify sanity checks for tesseroids
Make _check_tesseroid to run _longitud_continuity function. Redefine
what a valid tesseroid is and explain it on the doctstring of
_check_tesseroid. The _longitude_continuity function moves the
boundaries to the [-180, 180) interval and always copy the tesseroids
array. Any tesseroid with longitudinal boundaries outside [-180, 360]
are considered invalid. A tesseroid which longitudinal interval is
greater than one turn around the globe is also considered invalid.
Tests have been updated for these new standards.