Avoid copy of input arrays in forward modelling (#109)
Avoid casting the data type of input arrays in forward models. The dtype
argument is used only for building the arrays where the results are
being stored. 
For tesseroids, `_longitude_continuity` only copies the tesseroids array 
if needed and remove the numpy stacking of coordinates.
General improvements to how invalid tesseroids are caught and how 
longitude continuity is applied.

Fixes #100 and fixes #107