Wrap docstrings to 79 chars and check with flake8 (#127)
Wrap all docstrings to 79 characters of API functions. Exclude examples
(docstrings are rendered by sphinx-gallery) and test functions (not in public
APIs). Set `max-doc-length=79` characters to make flake8 check if any docstring
exceeds it. The setting checks comments as well and we found no way of
disabling that. So we'll format comments to 79 characters for consistency as
well. Minor modifications to the first line of some docstrings was required to
make them fit into a single line. Also made minor modifications on some
doctests to wrap to 79 characters.

Fixes #126