Replace global Zenodo DOI on badges and citation (#148)
Replace the reserved Zenodo DOI for the first release with the global
Zenodo DOI for Harmonica. Zenodo makes available the global DOI of the
release after it's released. So, the only DOI available at the time was
the one corresponding to the first release: v0.1.0. It's better to have
the global DOI on the badge from `README.rst` and on citation
instructions on `CITATION.rst`.