Grid on geodetic coordinates by passing projection
When gridding the gravity disturbance data we are doing so on a regular
grid defined on geodetic coordinates by passing the ellipsoid method
geodetic_to_spherical method as projection argument to the grid method
of EQLHarmonicSpherical. The regular grid is defined on geodetic
coordinates, but its coordinates are converted to spherical geocentric
before predicting. Use cartopy to generate the figures now that the grid
is defined on geodetic.