Start adding a base EQL class
Add new BaseEQL class as a subclass of vd.base.BaseGridder.
EQLHarmonic and EQLSphericalHarmonic are now subclasses of BaseEQL.
Redefined `grid`, `scatter` and `profile` methods to also ask for the
upward continue of the grid, scatter and profile points, respectively.
Pin verde to >=1.4.0 due to the change on the profile_coordinates
behaviour.