Improve EQLHarmonic classes
Add positional upward argument to grid, scatter and profile methods of
EQLHarmonic. Move EQLHarmonicSpherical to its own file and ditch the
projection argument on grid and scatter methods because it's
intended to be used with Cartesian gridders. Raise a NonImplementedError
on profile method for EQLHarmonicSpherical because we would need to
create the profile out of great-circle distances rather than absolute
differences of longitude and latitude angles. Add test functions and
update EQL examples.