Fetch a sample data before testing locate (#193)
After Pooch v1.1.1 the cache directory where all sample data are downloaded to
is created when a dataset is fetched and not on initialization as before. This
makes the test_datasets_locate() to fail on a fresh installation of Harmonica
(where no dataset has been downloaded yet). The solution proposed in Verde
(fatiando/verde#271) is to fetch one dataset before
running the locate() function.