Improve EQL harmonic classes (#190)
Make EQLHarmonic and EQLHarmonicSpherical independent subclasses of Verde's
BaseGridder. Add positional upward argument to grid and and profile
methods of EQLHarmonic. Raise a NotImplementedError when calling the
scatter method because it will be deprecated on Verde v2.0.0. Raise
a NotImplementedError when calling EQLHarmonicSpherical.profile method: the
profile should be created out of great-circle distances rather than absolute
differences of longitude and latitude angles. Remove the projection argument
on EQLHarmonicSpherical.grid method because it's intended to be used only on
Cartesian gridders. Reduce tests on EQLs with Numba disabled to speed them up.
Add test functions to catch NotImplementedErrors and to check if projection
is not a valid argument of EQLHarmonicSpherical.grid method.