Parallelize EQLs predictions and Jacobian build (#203)
Parallelize the build of the Jacobian matrix and the prediction step on
equivalent layer gridders through numba.prange iterators. This enables the
fit and predict methods to use all the available CPUs, greatly reducing the
computational time. The prange is only applied to the outer loops for each
task, i.e. distributing computation points between CPUs, but leaving the
iteration through the sources on single ones.