Optimize tesseroids forward model
Don't define a set of point masses to compute their effect, instead
compute results during runtime. This simplifies the code, reduces
memory consumption and prevents tesseroids from using
jit_point_mass_spherical, which is intended to be parallelized
afterwards. On tesseroids, parallelizing that step is not optimal: it's
better to parallelize the outer loops rather than the most inner ones.