Add function to create a layer of prisms (#186)
Add function that creates a layer of prisms as a xr.Dataset which contains
the coordinates of the center of each prism, its top and bottom boundaries (as
xarray coordinates) and each property (as density, magnetization, etc) as data
array. Add PrismsLayer class as a dataset accesor
for xr.Datasets, which allows to define several attributes, properties and
methods for the prism layer, which can be accessed through the prisms_layer
attribute of the xr.Dataset. Add a gravity method that computes
gravitational fields generated by every prism in the layer whose top boundary,
bottom boundary or density are not np.nans.
Add two gallery examples for the prisms layer, and a new South Africa
topography dataset, which is a downsample of ETOPO1.