Use the Jupyter book sphinx theme instead of RTD (#227)
The theme is a lot nicer and more modern. It has builtin a lot of the hacks we made 
to the template to add download and edit buttons. We could also do away with 
most of our custom templating and CSS tweaks. This opens up the possibility for 
further style improvements in the future (using admonitions, panels, etc). 
Notable changes: 
* the version number is now included in the page title
* the banner includes the slogan
* the front page disclaimer is redesigned and not gotten from the README anymore
* external links are included in `index.rst` instead of `conf.py` (this means no icons 
  and different titles but still looks nice)