Refactor the code for computing the derivatives
Move the derivatives.py to the root of the package.
Ditch the transform module and create a new filters module.
Add the functions that apply the directional derivatives in the
frequency domain to a new filters/derivatives.py file.
Create small wrappers for the xrft.fft and xrft.ifft functions inside
the filters module, although I'm planning to keep them private for now.
Move the coordinates shifting to the ifft wrapper function.