Run tesseroids in parallel by default
Enable parallelization through Numba prange on the computation points
loop. Both the stack and the small_tesseroids arrays are allocated
inside the same loop to avoid multiple threads to override the same
memory.
Add a new parallel argument to tesseroid_gravity and a new dispatcher
function that returns either the parallel or serial jitted function.