Allow prism_layer to take xarray objects as arguments (#243)
Extend the support of prism_layer to accept surface, reference, and
properties as xarray.DataArray. Improve the test functions by creating
a pytest fixture that parametrizes the data types of the output arrays:
either Numpy arrays or xarray.DataArrays. Generate a new pytest fixture
that returns a layer of prisms with some holes (missing prisms), ready to be
used as the expected results of some test functions. Parametrize the field on
each test function that computes gravitational fields of prism layers. Simplify
the prism_layer_topo example so the surface and the density of the layer are
passed as xarray objects.