Add block-averaged sources to EquivalentSources (#260)
Allow EquivalentSources to use the block-averaging strategy for locating the
sources introduced in Soler & Uieda (2021). Add a new block_size parameter to
define the size of the blocks. Add a new example to the Gallery and test
functions for the new feature. Add citation instructions for the paper when
using the block-averaged sources.