Add dtype argument to EquivalentSources
The dtype argument is used to cast the inputs of the fit method, to
build the location of the point sources, to allocate the Jacobian matrix
and therefore to produce predictions.
Add a new utility function for casting the inputs of the fit method to
the desired dtype.
Add new test file for the equivalent sources utility functions.
Add tests for the new feature.