Add variable density tesseroids (#269)
Add an implementation of the density-based discretization algorithm. Duplicate
the jit_tesseroid_gravity and the gauss_legendre_quadrature functions to
take a callable density argument. Reorganize the old and new tesseroid code,
moving some functions to private files. The dispatcher function chooses which
algorithm to use: constant density or variable density, serial or parallel. The
density-based discretization algorithm was modified to use the
scipy.optimize.minimize_scalar function to compute minimum and maximum
density values for normalization, and for computing the maximum absolute
difference. The density argument should be a function decorated with
numba.njit: we set this requirement to improve performance of the method. Add
example and test functions for the new feature. Add scipy as a dependency for
Harmonica.