Add gradient-boosted equivalent sources (#275)
Add a new EquivalentSourcesGB class that implements gradient-boosting
equivalent sources. It's a subclass of EquivalentSources, takes new
window_size and random_state arguments and has a modified fit method that
applies the gradient-boosting algorithm. The amount of overlapping is fixed to
50%. The windows are randomly shuffled by default. Add new test functions for
the new class and its features. Add a gallery example for gradient-boosted
equivalent sources.