Replace conda for pip on Actions and split requirements files (#282)
Create a new env directory with some requirements-*.txt files, each one
with the specific dependencies to run certain tasks, like checking style,
running the test, publishing to PyPI and building the docs. The GitHub Actions
now use pip to install the dependencies instead of having to setup Miniconda
and install them through conda. The only exception is the one that builds the
docs: we need pyproj and cartopy to do so and they are easier to install
through conda.