Ditch flake8-variables-names
In numerical computations we usually use single variable names, and
flake8-variables-names complains a lot about them (VNE001) so it would
be better to ignore it.
The VNE003 complains if we are using builtins as variable names, but
this is already covered by flake8-builtins.
The VNE002 doesn't introduce significant importance to our projects, in
fact we might use some of the words they complain about with good
reason.