Rename "point_mass_gravity" to "point_gravity" (#280)
Having gravity and mass in the function name is redundant: any point that
generates gravity must have a mass. The new name has the same structure as the
other forward modelling functions: tesseroid_gravity and prism_gravity.
Also renamed the source file to point_gravity.py, along with the tests and
the example. Added a point_mass_gravity() function for backward
compatibility: it wraps the point_gravity() function and raises
a FutureWarning.