Move tutorials to user_guide and use jupyter_sphinx
Instead of creating the user guide documents as .py scripts intended to
be run by Sphinx gallery example, now the user guide is a set of .rst
files that use the jupyter_sphinx extension for running code cells when
building the docs.
This eases the process of writing user guides and also organize them in
subtrees.