Add conversion of prisms or a prism layer to PyVista objects (#291)
Add a new visualization module that hosts prism_to_pyvista: a function to
convert a set of prisms into a pyvista.UnstructuredGrid. Include the new
module and this function in the API Reference. Add a new to_pyvista() method
to the PrismLayer accessor that converts a prism layer into a pyvista grid,
making it easier to plot it in 3D. The UnstructuredGrid has the information
about each prism as hexahedrons, along with their physical properties as cell
data. Add tests for the new features. Add pyvista and vtk as optional
dependencies to environment.yml and setup.cfg. Add a new example for
plotting a PrismLayer. Configure Sphinx to show pyvista plots in the gallery
and to use the pyvista-plot directive in docstrings.