Add deprecations to datasets and synthetic modules (#304)
Add FutureWarnings to public functions of the synthetic and dataset
modules. Add tests for the new warnings. Both modules will be deprecated in
Harmonica v0.6.0. Instead of providing sample datasets, Harmonica will depend
on Ensaio for that. The synthetic surveys depend on some of the sample
datasets, but those functions are intended to be used in methodology articles,
so they should live somewhere else.