Ditch soon-to-be deprecated args of equivalent sources grid method (#311)
The grid() method of Verde gridders now take a coordinates argument with
the coordinates of the target grid. The previous region, shape and
spacing arguments will be deprecated in Verde v2.0.0. This change makes it
easier for our equivalent sources classes: we don't need the extra upward
argument, users can create the coordinates of the target grid using
verde.grid_coordinates and pass them via coordinates argument. Ditch the
upward, shape, spacing and region arguments from the equivalent sources
gridders. Replace them for the new coordinates argument: users need to
provide the coordinates of the target grid instead of building it through the
grid method. Raise errors if any of those old arguments are being passed. Raise
warnings if any kwargs are passed: they are being ignored and not passed to the
BaseGridder.grid() method.