Add function to compute Moho depth based on Airy isostasy hypothesis (#307)
Create a new `isostatic_moho_airy` function that computes the depth of the Moho
based on the Airy isostasy hypothesis. It takes the height of the crystalline
basement and an optional dictionary of additional layers on top of it. Each
layer is specified by its thickness and density. The function makes use of the
_rock equivalent topography_ concept to compute the depth of the Moho taking
into account. Add `FutureWarning` to `isostasy_airy` function warning of
deprecation after next release. Add tests and a new example for the new
function. Ditch the example for `isostasy_airy`.