Fix test function for empty ICGEM gdf file (#345)
The test for checking if proper errors and warnings were raised when reading an empty ICGEM gdf file was failing in my local system due to a change in the Numpy warning for the empty file. Took the opportunity to refactor the test using pytest.raises and pytest.warns to capture the errors and warnings.