Add function to create a tesseroid layer (#316)
Add function that creates a tesseroid layer, similar to the prism layer. Create
an Xarray DataAccesor that makes available some methods to deal with the prism
layer, included one to compute the gravitational field generated by the whole
layer in a set of computation points. Add checks to avoid generating layers
with tesseroids that overlap around the earth. Add tests and a new gallery
example.