Set lower bounds for supported dependency versions (#356)
Follow NEP29 and set lower bounds for the dependencies (all minor versions from the last 2 years). The only exception is xrft, which is new and likely requires v1.0 to work with Harmonica. Setup CI to run on lowest Python with lowest dependencies, highest Python with highest dependencies, and highest Python with optional dependencies. Use Dependente to extract the version information from setup.cfg instead of the custom export tool we had.