Update README to match Verde and Boule (#358)
Remove the duplicated heading saying "Harmonica" and remove the slogan from the
banner since it's in the text already. Remove the GitHub Actions badge which
uses a third-party service and Actions status is easily seen on GitHub anyway.
Add a conda-forge badge for easily finding the feedstock. Update contact
information. Move disclaimer to a "Project status" section.

Co-authored-by: Santiago Soler <santiago.r.soler@gmail.com>
Co-authored-by: Santiago Soler <santisoler@fastmail.com>