Configure flake8 to ignore spaces before a colon
Flake8 complains with leaving a space before a color, what makes sense
for colons that go after a statement (function definition, for loop, if,
etc). But it complains about the spaces before a colon used in slicing.
Black already handles the first case, so it makes sense to ignore E203.