Add function to read Oasis Montaj© grd files (#348)
Add a new `load_oasis_montaj_grid` function that reads Geosoft Oasis Montaj©
GRD files and returns an `xarray.DataArray` that includes the 2D grid values,
its coordinates and the information available in the file header. The function
supports reading uncompressed and compressed grids (with gzip); rotated and
unrotated grids; orders of +/- 1 (the order in which the grid values are stored
in the file); every possible data type supported in GRD files. The function
doesn't support orders besides +/- 1, colour grids nor grids. Compressed grids
are assumed to be compressed with gzip regadless that the header specifies that
it uses LZRW1. The new function lives in a newly created `_io` submodule, where
the `load_icgem_gdf` function has been moved to.

The support for compressed grids was inspired by the code in
https://github.com/Loop3D/geosoft_grid copyrighted by Loop3D and released under
the MIT License.

References:
- https://help.seequent.com/Oasis-montaj/9.9/en/Content/ss/glossary/grid_file_format__grd.htm

Co-authored-by: LL-Geo <54405391+LL-Geo@users.noreply.github.com> and
mdtanker <81199856+mdtanker@users.noreply.github.com>