Refactor reduction to the pole and pseudogravity filters
Move some portions of the RTP code into its own private function that
both the RTP filter and the pseudogravity filter can use. Refactor the
code to generate the filter to improve readability and make it a little
bit more optimal by predefining the cosines and sines. Define a private
function to check if the two angles for the magnetization are both None
or both angles but not one None and one angle.