Add option to discard thin prisms of a layer when forward modelling (#373)
Add a new `thickness_threshold` parameter to the `prism_layer.gravity()` method
that allows the user to ignore prisms below a certain thickness. This is
implemented within a new `_discard_thin_prisms()`, function which is similar to
the `_discard_null_prisms()` function. Add tests that check whether the
calculated gravity is the same from manually removing prisms and removing
prisms based on a threshold, and whether providing no threshold still produces
the correct gravity.

Co-authored-by: Santiago Soler <santisoler@fastmail.com>