Clean API making some submodules private (#362)
Make the `forward`, `equivalent_sources`, `gravity_corrections`,
`isostasy` and `transformations` submodules as private ones. This will
enforce the usage of public functions and classes allowing developers to
refactor private ones without the risk of braking backward
compatibility. It also cleans up the list of options available when
using autocompletion in IDEs and JupyterLab, helping the discoverability
of the functions and classes we have in Harmonica.