Drop null prisms when converting a prism layer to pyvista (#393)
Add an optional `drop_null_prisms` flag to the prism layer
`to_pyvista()` method that ditches the prisms that have zero volume or
`nan` values in their top or bottom boundaries. These null prisms won't
be included in the PyVista Unstructured grid. By default the flag is set
to `True`. Add a new test for this new feature.