Add easting and northing components of acceleration
Allow the `prism_gravity` function to compute the easting and northing
components of the gravitational acceleration of prisms, using Choclo
kernels for it. Add tests that compare with bare Choclo results. Remove
the dispatcher and replace it with and `if`/`else` statement for
selecting the parallelized or serialized jitted function.