Warn for prism gravity tensor on singular points
Check if the gravity tensor components of a prism are being computed at
any singular point. Choclo will return a ``np.nan`` on those points, but
it's worth warning the user about it before they get the results.
Add tests that check if the warnings are being raised on those cases.