Update environment.yml and some dependencies
Update `environment.yml`: remove some pinned versions of dependencies
that aren't critical, install `numba-progress` from conda-forge, unpin
pygmt, gmt and gdal. Remove `pandas` from the list of required
dependencies and move it to the ones needed to build the docs. We don't
use pandas directly in Harmonica, but through Verde, which is already
a Harmonica dependency.