Use Dependabot to update GitHub Actions workflows
The Actions we use have pinned versions and it's almost impossible to
keep up with new releases of them. Add configuration for Dependabot to
track the Actions versions and send pull requests when new versions are
available. This only works if version numbers are pinned instead of the
commit hashes. I think the benefit of being updated out weight any risk
associated with this.