Use Dependabot to update GitHub Actions workflows (#455)
The Actions we use have pinned versions and it's almost impossible to
keep up with new releases of them. Add configuration for Dependabot to
track the Actions versions and send pull requests when new versions are
available. Remove the Cancel Workflow Action to avoid giving full access
to a third party action.

Co-authored-by: Santiago Soler <santisoler@fastmail.com>