Fix inverted sign in upward derivative filter (#479)
Since the derivative is in the upward direction, the fields should get smaller 
(negative derivative). This is is the opposite of what we had because of a 
missing minus sign in the filter. The tests didn't fail because we compared 
the upward derivative of the gravity potential against g_z, which is positive 
downward. Fix the tests as well to compare against g_up instead of g_z.