Add job for pushing coverage to codecov in Actions
Remove the push to codecov step from the `test` job into a new job that
depends on the test job. Upload the coverage report as an artifact after
testing, and reuse that artifact in the new job. Only the coverage from
the ubuntu runner with the latest dependencies is uploaded to minimize
hits to codecov.