Speed up for a factor of 2
Allocate the output array out of the loop. This way we can do it once
only when evaluating the spherical harmonics. Pre-compute the square
roots of integers that we use in the loops. Doesn't work well if the
square roots are calculated at the module level (probably a numba
thing).