Change default for window size in EquivalentSourcesGB (#487)
Change the default value for the `window_size` in the
`EquivalentSourcesGB` constructor to `"default"`: window size will be
estimated to locate approximately 5000 data points in each window. If
`window_size="default"` and less than 5000 data points are used to fit
the sources, a single window for all data and sources will be used.
Check if the passed value of `window_size` is valid. Add a new
`window_size_` attribute to the class that is set after fitting the
sources. Update docstrings and extend tests.

---------

Co-authored-by: Santiago Soler <santisoler@fastmail.com>