Add associated Legendre function calculations (#505)
The numba-based functions calculate the associated Legendre functions
(unnormalized, Schmidt normalized, and fully normalized) and their
derivatives $\dfrac{\partial P_n^m}{\partial \theta}(\cos\theta)$. The
derivatives functions can be used to calculate higher order derivatives
as well. Values are tested against analytical solutions for the first 4
degrees as well as an identity for the Schmidt functions and the
Legendre equation for the Schmidt and fully normalized functions. The
values are returned as 2D numpy arrays. This is a bit of a waste of
memory but it's much easier to use than storing these in sparse matrices
or 1D arrays. It should be acceptable since we're unlikely to ever go over 
~2800 degree so the waste is in the order of a few Mb at worse.

Co-authored-by: Santiago Soler <santisoler@fastmail.com>