Change default value of depth in equivalent sources (#491)
Change the default value of the `depth` argument in `EquivalentSources`
to `"default"`. If this value is passed, then the depth of the sources
is set as 4.5 times the mean distance between first neighbouring
sources. If a numerical value is passed, this is the one that will be
used. Introduce a new `depth_` attribute where the estimated/passed
numeric value is stored. If `points` is passed, the `depth_` attribute
is set to None. Add a test checking that the depth value is properly
set. Merge the tests functions that checked the accuracy of equivalent
sources using different dtypes.

---------

Co-authored-by: Leonardo Uieda <leo@uieda.com>