Fix equiv sources tests failing after sklearn 1.6.0
Test cartesian equivalent sources with f32 and f64 dtypes and also with
`damping` as `None` and with a very small value. Mark the combination of
`damping=None` and `dtype=np.float32` with `xfail`, so it gets flagged
but `pytest` doesn't errors out. Since the equivalent sources with
damping are more accurate, we don't need to have different absolute
tolerances for both cases: keep a single value for `atol`. Set some very
low absolute tolerances in spherical equivalent sources tests.