Fix vector to angles for small intensities (#565)
When the intensity is very small (<1e-10), the vector to angles
calculation fails because of the masking that was used. It rounds these
numbers to zero and returns always inc=90, dec=0. To handle this, use
the arctangent instead of arcsine in the equations. This was, we can use
arctan2 which is able to handle 0 in the denominator. Add tests to make
sure this doesn't happen again. These small values appear in magnetic
microscopy and we ran into this issue in Magali.