Simplify code for gravity fields of ellipsoids
Move function to compute the elliptical integrals A(lambda), B(lambda),
and C(lambda) to `utils_ellipsoids.py` and rename it to
`get_elliptical_integrals`, since they are going to be reused by the
magnetic function. Define a single `_compute_gravity_ellipsoid` function
in `ellipsoid_gravity.py` that computes internal and external gravity
fields of a single ellipsoid in the local coordinate system, for any
type of ellipsoid. This function makes use of the new
`get_elliptical_integrals` function in `utils_ellipsoids.py`. Minor
cleanup of the `ellipsoid_gravity` function.