Add IGRF forward calculation (#556)
Add a class to calculate the IGRF14 field. Uses Pooch to download IGRF14
from Zenodo. The class takes a date to instantiate and then calculates
the field for that data through `predict` and `grid` methods. Inputs and
outputs are in geocentric spherical coordinates by default but can be
changed to geodetic if a Boule ellipsoid is given to the constructor.

Co-authored-by: Matt Tankersley <81199856+mdtanker@users.noreply.github.com>