Reorganize the magnetic code a bit
Create a private function for computing the magnetic field of a single
ellipsoid, to declutter the public function and reduce nested levels.
Avoid building a mask for internal points since we are already iterating
over the observation points with a regular for loop. Fix definition of
the h field for internal points: it doesn't include the magnetization,
the b field does. Add a few TODOs for later.