Keep cleaning up magnetic code
Improve the sanity checks for the inputs of the public function. Move
the casting of the remanence to the function that computes the magnetic
field of a single ellipsoid, so we can pass an heterogeneous list as the
remanence (including Nones). Rename the `calculate_lambda` function to
make it public in the module (but not exposed in the API).
Add test for multiple ellipsoids with remanence.