Optimize lambda for prolate and oblate ellipsoids (#604)
Solve the second order equation for lambda when we are working with
prolate and oblate ellipsoids. The second order equation doesn't require
to compute trigonometric functions. Extend tests to arrays, and only on
external points. Move `is_internal` to ellipsoid's utils. Calculate
lambda only on external points in the gravity forward. Move the
computation of lambda to the private function that computes the
demagnetization tensor on external points: we don't use lambda anywhere
else.