Add a new `Sphere` class as a special case of ellipsoids (#602)
Add a new `Sphere` class as a special case of ellipsoids with `a == b ==
c`. Implement gravity and magnetic forward modelling for the sphere,
both inside and outside of the body. Extend tests to `Sphere` objects.
Use the forward modelling of the `Sphere` to compare the fields of
ellipsoids against it.