Add a new `create_ellipsoid` function (#606)
Add a new `create_ellipsoid` function that takes values of the three
semiaxes lenghts (`a`, `b`, and `c`) in any particular order and returns
an ellipsoid object of the corresponding type. It takes the `center` as
required argument as well, but defaults yaw, pitch, and roll angles to
zero. It can also take physical properties of the ellipsoids that are
passed through ``kwargs``. Add tests for the new function, and
reorganize the ellipsoid code and tests under new `ellipsoids`
submodule.

**Relevant issues/PRs:**

Part of #594